Fixed pagination button click zone
Line 76 by using .firstElementChild I constrained the event listener to the button (not its parent <li> which is why clicking in between buttons was triggering event before.

I noticed in the Console an Uncaught Type error in the no results condition. I think that's because linkList is empty so there's nothing to add active class to.
Need to fix this.